Move layouts outside of the graphs themselves. Standardize the way layouts are sent and received for a graph. Still needs ability to read from JSON, pass in layout, and documentation.
